+++ src/ZubZet.php
@@ -9,10 +9,12 @@
     use ZubZet\Framework\Message\Response;
     use ZubZet\Framework\Authentication\User;
     use ZubZet\Framework\Database\Connection;
-    use ZubZet\Framework\Message\Input\State as Input;
+    use ZubZet\Framework\Resources\AssetProxy;
     use ZubZet\Framework\Core\CanRetrieveModel;
     use ZubZet\Framework\Bootstrap\Configuration;
     use ZubZet\Framework\Support\GlobalReferences;
+    use ZubZet\Framework\Message\Input\State as Input;
+    use ZubZet\Framework\Core\CanRetrieveBooterSettings;
     use ZubZet\Framework\ErrorHandling\ExceptionBehavior;
 
     class ZubZet {
@@ -20,6 +22,7 @@
         use Configuration;
         use CanRetrieveModel;
         use ExceptionBehavior;
+        use CanRetrieveBooterSettings;
 
         /** @var Connection $z_db Database proxy object  */
         public $z_db;
@@ -42,6 +45,12 @@
         public static ?ZubZet $instance = null;
 
         /**
+         * @internal
+         * @var AssetProxy The instance of the asset proxy
+         */
+        public AssetProxy $assetProxy;
+
+        /**
          * Parses all the options as variables, instantiates the z_db, and establishes the db connection.
          */
         function __construct(array $params = []) {
@@ -55,6 +64,8 @@
 
             // Error handling
             $this->setExceptionBehavior();
+
+            $this->assetProxy = new AssetProxy;
 
             // Static imports
             new Constants;